Initial public commit.
Program works as is right now.  Currently working on implementing programs so that program will automate a fermentation process.  This commit has a few sections commented out so that you can run it and just see the interface.  To get the full functionality, you will need to uncomment the sections "MARK: Comment / uncomment to test"